pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "supermd"
version = "2.5.0"
description = "Supernote to Markdown converter and sync tool"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    "llm",
    "llm-ollama",
    "Jinja2",
    "numpy",
    "pillow",
    "pydantic>=2.0",
    "PyMuPDF",
    "pypng",
    "PyYAML",
    "tqdm",
    "click",
    "watchdog",
    # Transitive dependencies for vendored libs
    "platformdirs", # supermd
    "potracer",     # supernotelib
    "svgwrite",     # supernotelib
    "colour",       # supernotelib
    "svglib",       # supernotelib
]

[project.optional-dependencies]
test = [
    "pytest",
]

[project.scripts]
supermd = "supermd.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
